% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{estimate_daylength}
\alias{estimate_daylength}
\title{Estimate Astronomical Day Length (Photoperiod)}
\usage{
estimate_daylength(lat, doy)
}
\arguments{
\item{lat}{Latitude in decimal degrees (negative for South).}

\item{doy}{A numeric vector of day of the year (Julian day, 1 to 365/366) or
a vector of \code{Date} / \code{POSIXt} objects.}
}
\value{
A numeric vector of day length values in decimal hours.
}
\description{
Calculates the astronomical day length (photoperiod) in decimal hours.
}
\examples{
estimate_daylength(lat = -27.28, doy = 172) # Winter solstice
estimate_daylength(lat = -27.28, doy = 355) # Summer solstice
}
