% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{estimate_solar_noon}
\alias{estimate_solar_noon}
\title{Estimate Solar Noon}
\usage{
estimate_solar_noon(doy, lon = NULL, tz = NULL)
}
\arguments{
\item{doy}{A numeric vector of day of the year (Julian day, 1 to 365/366) or
a vector of \code{Date} / \code{POSIXt} objects.}

\item{lon}{Optional. Longitude in decimal degrees (negative for West).
If omitted, local solar noon is returned (12.0).}

\item{tz}{Optional. Time zone offset from UTC in hours (e.g., -3 for UTC-3).}
}
\value{
A numeric vector representing solar noon in decimal hours (0 to 24).
}
\description{
Calculates the solar noon time in decimal hours. By default, it returns
local solar noon (12.0 h), but it can compute civil standard clock time
when longitude and time zone offset are provided.
}
\examples{
# Local solar noon (default)
estimate_solar_noon(doy = 180)

# Civil clock time for solar noon with longitude and UTC offset
estimate_solar_noon(doy = 180, lon = -50.58, tz = -3)
}
