% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{estimate_sunset}
\alias{estimate_sunset}
\title{Estimate Sunset Time}
\usage{
estimate_sunset(lat, doy, solar_noon = 12)
}
\arguments{
\item{lat}{Latitude in decimal degrees (negative for South).}

\item{doy}{A numeric vector of day of the year (Julian day, 1 to 365/366) or
a vector of \code{Date} / \code{POSIXt} objects.}

\item{solar_noon}{Optional. Numeric vector or scalar indicating solar noon
in decimal hours. Default is 12.0.}
}
\value{
A numeric vector containing sunset times in decimal hours.
}
\description{
Computes sunset time in decimal hours based on latitude, day of the year,
and solar noon.
}
\examples{
estimate_sunset(lat = -27.28, doy = 355)
}
