% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{estimate_hourly_rh}
\alias{estimate_hourly_rh}
\title{Estimate Hourly Relative Humidity from Daily Metrics}
\usage{
estimate_hourly_rh(rh_daily, t_daily, t_hourly)
}
\arguments{
\item{rh_daily}{A single numeric value representing the daily average relative humidity (\\%).}

\item{t_daily}{A single numeric value representing the daily average temperature (°C).}

\item{t_hourly}{A numeric vector of length 24 containing the hourly temperatures (°C).}
}
\value{
A numeric vector of length 24 containing estimated hourly relative humidity values (\\%).
}
\description{
Reconstructs a 24-hour profile of hourly relative humidity values
based on daily average relative humidity, daily average temperature, and a
vector of hourly temperatures. It assumes that the actual vapor pressure
remains constant throughout the day.
}
\examples{
daily_rh <- 80
daily_t <- 20
hourly_t <- c(16, 15, 14, 14, 15, 17, 19, 21, 23, 24, 25, 25,
              24, 23, 22, 21, 20, 19, 18, 17, 17, 16, 16, 16)

estimate_hourly_rh(rh_daily = daily_rh, t_daily = daily_t, t_hourly = hourly_t)
}
